requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-chroma
gradio
huggingface_hub
python-dotenv
